app/requirements.txt
Flask>=3.0,<4.0
gunicorn>=22,<23
google-cloud-bigquery>=3.18,<4.0
psycopg2-binary>=2.9,<3.0      # opcional si finalmente consultas Postgres
